+++ src/Tests/MeasureMap.UnitTest/ExecutionContextTests.cs
@@ -75,5 +75,23 @@ namespace MeasureMap.UnitTest
 
            context.Get<int>("SomeKey").Should().Be(4);
         }
+        
+        [Test]
+        public void ExecutionContext_Get_NullValue()
+        {
+            var context = new ExecutionContext();
+            context.SessionData.Add("somekey", null);
+
+            context.Get<int>("SomeKey").Should().Be(0);
+        }
+        
+        [Test]
+        public void ExecutionContext_Get_InvalidType()
+        {
+            var context = new ExecutionContext();
+            context.SessionData.Add("somekey", "test");
+
+            context.Get<int>("SomeKey").Should().Be(0);
+        }
     }
 }